Renamed conversion subfolders by their Dandiset ID.
